applied all content for the week ready for merge
