+++ pages/api/auth/[...nextauth].js
@@ -1,0 +1,123 @@
+import NextAuth from "next-auth";
+import { MongoDBAdapter } from "@next-auth/mongodb-adapter";
+import clientPromise from "../../../lib/mongodb";
+import SpotifyProvider from "next-auth/providers/spotify";
+// import spotifyApi, { LOGIN_URL } from "../../../lib/spotify";
+
+const SPOTIFY_AUTHORIZATION_URL =
+  "https://accounts.spotify.com/authorize?" +
+  new URLSearchParams({
+    prompt: "consent",
+    access_type: "offline",
+    response_type: "code",
+  });
+
+/**
+ * Takes a token, and returns a new token with updated
+ * `accessToken` and `accessTokenExpires`. If an error occurs,
+ * returns the old token and an error property
+ */
+async function refreshAccessToken(token) {
+  try {
+    // spotifyApi.setAccessToken(token.accessToken);
+    // spotifyApi.setRefreshToken(token.refreshToken);
+
+    // const { body: refreshedToken } = await spotifyApi.refreshAccessToken();
+    // console.log("Refreshed token is", refreshedToken);
+
+    const url =
+      "https://accounts.spotify.com/api/token?" +
+      new URLSearchParams({
+        client_id: process.env.SPOTIFY_CLIENT_ID,
+        client_secret: process.env.SPOTIFY_CLIENT_SECRET,
+        grant_type: "refresh_token",
+        refresh_token: token.refreshToken,
+      });
+
+    const response = await fetch(url, {
+      headers: {
+        "Content-Type": "application/x-www-form-urlencoded",
+      },
+      method: "POST",
+    });
+
+    const refreshedTokens = await response.json();
+
+    if (!response.ok) {
+      throw refreshedTokens;
+    }
+
+    return {
+      ...token,
+      accessToken: refreshedTokens.access_token,
+      accessTokenExpires: Date.now() + refreshedTokens.expires_in * 1000,
+      refreshToken: refreshedTokens.refresh_token ?? token.refreshToken, // Fall back to old refresh token
+    };
+  } catch (error) {
+    console.log(error);
+
+    return {
+      ...token,
+      error: "RefreshAccessTokenError",
+    };
+  }
+}
+
+export default NextAuth({
+  adapter: MongoDBAdapter(clientPromise),
+  secret: process.env.NEXTAUTH_SECRET,
+  providers: [
+    SpotifyProvider({
+      clientId: process.env.SPOTIFY_CLIENT_ID,
+      clientSecret: process.env.SPOTIFY_CLIENT_SECRET,
+      authorization: SPOTIFY_AUTHORIZATION_URL,
+    }),
+  ],
+  database: process.env.MONGODB_URI,
+  session: {
+    strategy: "jwt",
+  },
+  secret: process.env.NEXTAUTH_SECRET,
+  jwt: {
+    secret: process.env.NEXTAUTH_JWT_SECRET,
+  },
+  callbacks: {
+    async jwt({ token, account, user }) {
+      // initail sign in
+      if (account && user) {
+        return {
+          ...token,
+          accessToken: account.access_token,
+          refreshToken: account.refresh_token,
+          username: account.providerAccountId,
+          accessTokenExpires: account.expires_at * 1000, // we are handling expiry times in milliseconds hence * 1000
+          user,
+        };
+      }
+
+      // Return previous token if the access token has not expired yet
+      if (Date.now() < token.accessTokenExpires) {
+        console.log("EXISTING ACCESS TOKEN IS VALID");
+        return token;
+      }
+
+      // Access token has expired, so we need to refresh it ...
+      console.log("ACCESS TOKEN HAS EXPIRED, REFRESHING...");
+      return await refreshAccessToken(token);
+    },
+
+    async session({ session, token }) {
+      session.user = token.user;
+      session.accessToken = token.accessToken;
+      session.error = token.error;
+
+      return session;
+    },
+  },
+
+  // secret: process.env.NEXTAUTH_SECRET,
+  // pages: {
+  //   signIn: "/auth/signin",
+  //   signOut: "/"
+  // }
+});
