+++ styles/globals.css
@@ -18,3 +18,8 @@ a {
 * {
   box-sizing: border-box;
 }
+
+.logo{
+  width: 50px;
+  height: 50px;
+}
